Operator pulls from addon preferences "vertex count" for default initial number of segments.
Note that for non closed forms (eg, not tubes), it will specify the number of segments not the
number of vertices  (N_verts -1).  However this is a pretty trivial point.  We may want to
change the variable name for consistency or change the behavior to force the number of
verts instead of number of segments.