temporary hot key 'S' to snap the scene 3d cursor to the COM of the selected cut line
improves issue #73 allowing users to lock scene view to cursor.  However this is sub ideal
because in this mode, the user can no longer pan around.  This is probably a decent compromise

Improves issue #69 Initial implementation oft edge slide widget behavio.
It currently only modifies the cut plane location and does not smoothly interpolate the
cut directions between the adjacent cuts.

It will be impossible to have the ring verts "track" to existing edges because when the contour is
recut, it derives it's initial vertex spacing from the underlying topology which by definition is going to
be unpredictable.  Users should use the contour tools "edge slide" when changing a cut
that will cut across a mesh where edge sliding between loops would not produce satisfactory results.  Eg,
if there are shape changes being captured in the new loop position not adequately interpolated
by two exsiting loops.