Smoothing of the cut directions function added to CutSeries class
Average of the cut directions added to the CutSeries class allowing all the cuts to be parallel

Function to asses a new cut for inclusion into the series.  Criteria is more strict than when trying
to build a series out of a collection of cuts that is agnostic to the order they were derived.
However, because we have a user stroke which gives us information about the order of cuts.

Dummy event keys added to the modal for testing.