Add polyfill for nicer face visualization.
This way the faces between loops now draw with 20% alpha of the current geometry theme color.

Note: I've added the draw_quads_from_3dpoints to the Contour Utilities, but this should really reference common_drawing. Will refactor later along with the rest of the codebase, such that all utilities are pulling from the common lib.